align tests directory structure to project one
