+++ .ci/docker-compose.yml
@@ -8,6 +8,7 @@ services:
       target: test
     environment:
       - DB_STRING=postgresql://postgres:test@ci_db:5432/postgres
+      - SQLALCHEMY_SILENCE_UBER_WARNING=1
 
   integration:
     depends_on:
@@ -18,6 +19,7 @@ services:
       target: test
     environment:
       - DB_STRING=postgresql://postgres:test@ci_db:5432/postgres
+      - SQLALCHEMY_SILENCE_UBER_WARNING=1
 
   ci_db:
     image: postgres:14.5
