+++ alembic.ini
@@ -1,6 +1,6 @@
 
 [alembic]
-script_location = api/migrations
+script_location = api/infrastructure/database/migrations
 prepend_sys_path = .
 version_path_separator = os  # Use os.pathsep. Default configuration used for new projects.
 sqlalchemy.url = driver://user:pass@localhost/dbname
