+++ api/tests/unit/usecases/test_get_starneighbours.py
@@ -1,0 +1,5 @@
+import pytest # type: ignore
+
+@pytest.mark.unit
+def test_get_starneighbours():
+    assert 1 != 0
