+++ api/tests/unit/usecases/test_login.py
@@ -1,0 +1,5 @@
+import pytest # type: ignore
+
+@pytest.mark.unit
+def test_login():
+    assert 1 != 0
